WIP: fetch-pack: warn when mapping is required but unsupported
When we have or are requesting a shallow clone and we have a
compatibility algorithm, we require the remote side to support both of
our algorithms.  This is because shallow clones provide us an incomplete
history and without help from the server, we cannot map all of the
objects.

In such a case, let's provide a more helpful response than an "unmapped
shallow" error message, which most users would not understand.

WIP: move after v0 version
WIP: include t1017 patch

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>