shallow: read and write shallows in proper algorithm
We currently have support for fetching a repository in the client's
compatibility algorithm.  When we're performing such an operation, we
need to convert the object IDs used in shallows to the correct algorithm
before sending them over the protocol.

Update the write_shallow_commits function to take an algorithm to use
and if one is provided, convert it to the desired algorithm.  Store the
algorithm in the callback data so that we can use it for formatting as
we iterate over the shallows. Change the callers which are not in the
fetch-pack code to use the repository's main algorithm as part of this.

Make sure to also set the packet reader's hash algorithm since we use
that to format and parse some object IDs in our communications with the
other side.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>