rust: don't expose the internals of ObjectID
It's tidier to not expose the internals of ObjectID and it also allows
us to gracefully make more assumptions that the internals are in a
reasonable state.  Update all callers to use the existing accessors or
constructors and update one caller to write the code in an approach
using iterators now that that's possible.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>