loose: pass object type to for-each function
In a future commit, we'll want to repack the loose object maps, which
means we'll need to iterate over each object and find its two object
IDs as well as the type of the object.  Expand the iterator function to
provide the type as well as the object IDs and add a test for the
iterator as well.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>