loose: expose loose object map iterator
In a future commit, we'll repack several of our loose object maps into
one consolidated one.  In order to do that, we'll need to have access to
the items in our existing maps, so expose the iterator function so that
we can use it in the GC code.

Even though the for-each function is not used outside of a
Rust-containing build, expose it more generally so it can be used in
prototypes of functions which are agnostic to whether the Rust code is
compiled.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>