gc: repack loose object maps
When we add loose objects or other objects that require mapping into the
repository and we're in compatibility, we write loose object maps into
the repository.  However, like with packs, too many loose object maps
can negatively impact performance.

Add functionality git gc to repack the loose object maps.  Add a
configuration option to control how often to run an auto GC in this
case.  Include this functionality inline since we don't have a separate
command to handle these maps.

When repacking the loose object maps, do not include objects we have
repacked or special objects (the empty tree, empty blob, and null OID).
This means that if there are no unreachable objects, no loose object map
will be written since the objects will all be packed.  However, add a
test with unreachable objects to ensure that we do correctly write them
nonetheless.

In some cases, one of the loose object maps we've previously created is
the same as the new one we've written.  Be careful not to delete that
one, since that would cause us to lose data.  Our test covers this case
since there is only one loose object being preserved.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>