loose: expose location of written map
Right now, the location of the map we've written is not exposed
anywhere.  So far, this has been fine, but when we're repacking the
loose object maps, we'll need to know the name of the file we've just
created so that we don't accidentally remove it.  Expose this
information with an optional file parameter, which can be set to NULL if
the caller is not interested in the value.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>