t3800: use BROKEN_OBJECTS prerequisite
There are several tests here that attempt to test the behavior of broken
tags.  However, when we're in interoperability mode, it is impossible to
create broken objects because they cannot be converted into the other
algorithm.  Use the BROKEN_OBJECTS prerequisite to skip tests which
cannot work and check for an appropriate error message.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>