upload-pack: make object-map-info work for protocols v2 and v0
Right now, the code to write map-object lines is specific to protocol v2
because it uses section headers and delimiter packets.  However, it can
easily be changed to be generic and useful for protocol v0 as well by
simply omitting these two features.  Make them conditional on a boolean
variable and set that to true in our v2 code in preparation for our
reuse of this code in protocol v0.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>