upload-pack: support shallow fetches with protocol v0
We already have shallow fetch and clone support with protocol v2, but
there are some cases where people still use protocol v0.  Accept an
object-format-map capability with which the server can advertise an
additional algorithm, and if the client also supports this capability,
send an object mapping in that algorithm for shallows, their parents,
and their trees.

We need all three components because the server sends the commits in the
shallows and they cannot be appropriately mapped without the trees and
parents.  For efficiency, simplicity, and security reasons, we want to
convert all objects we receive ourselves instead of trying to
special-case certain objects.

Send the mappings immediately before the shallows so that we can process
items in a single pass.  We will add additional types of mappings in the
future as well which will be sent in the same way.

Allow the server to control this mapping with a configuration option
which defaults to on.  There is little functional downside to offering
it by default and doing so makes much of our testsuite start passing in
compatibility mode.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>