fetch-pack: support shallow fetches with protocol v0
We already have shallow fetch and clone support with protocol v2, but
there are some cases where people still use protocol v0.  Accept an
object-format-map capability from the server and, if we also support the
algorithm, send the capability ourselves.

Then, read the map-object lines, if any, as well as shallow and
unshallow lines in the correct algorithm, and save the mappings into the
loose object map.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>