shallow: allow advertising shallows in compat algorithm
Right now, when we push, the server advertises any relevant shallows
after the ref advertisement.  In the future, we will want to support
transfers with multiple algorithms, in which case we'll need to send
directives to map the shallow objects, their parents, and their trees.
The parents and trees of the shallow entries must be sent in order to
have the pack code map the shallow commit itself properly.

Send directives to map the shallows over the wire when we're advertising
the refs by first collecting the shallows, then advertising the mapping,
then the shallows.  This allows one-pass processing.

De-duplicate the same code we have in send-pack as well but which just
writes data to a strbuf by always writing to the strbuf and then just
sending it over the wire when complete.  Expose the necessary function
in the headers since it now lives in shallow.c.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>