transport: move code to parse map-object lines
We have code in fetch-pack to process map-object lines, but this is also
useful for send-pack as well.  Move this code into the transport code more
generally and expose it there.

While we're at it, be sure to pass in a repository argument to make
working with different repositories easier.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>