commit: allow parsing parents of shallow objects
When we're sending the shallows over the protocol in interoperability
mode, we need to also send the parents and the tree in order to be able
to successfully map the object correctly.  This means that index-pack
doesn't need to learn to special case the shallow points.

In order to make this possible when we're pushing into a shallow
repository, we'll need a way to find out what the parents of a shallow
commit are even though we don't normally want to have them.  Add the
ability to invoke repo_parse_commit_gently with an extra argument to
keep these parents for this case.

Signed-off-by: brian m. carlson <bk2204@github.com>