shallow: specify both algorithms for mapping shallow object IDs
In interoperability mode, we need to advertise a mapping for shallow
objects in both the protocol algorithm and the other algorithm (which
need not be the main and compatibility algorithm, respectively).  Adjust
our shallow code to take the protocol and mapping algorithm and map each
shallow, including its parents and tree.  Re-parse the object if
necessary to get its parents without affecting other users.

In several cases, add a variable for the mapped algorithm which is set
to NULL.  In future commits, we'll wire these up to handle the correct
algorithm by initializing it to an appropriate value, but right now this
preserves the status quo by sending no mapping.

Signed-off-by: brian m. carlson <bk2204@github.com>