shallow: convert object IDs when processing shallows
When we're processing the shallows during receive-pack, we need to be
sure to convert the object IDs to the main algorithm before asking
whether we have an object or before we look up a commit.  Otherwise, we
may incorrectly determine that we don't have the object ID and fail to
add or prune the expected object from the list of shallows.

Signed-off-by: brian m. carlson <bk2204@github.com>