Implement pushing for shallow repositories in interop mode
Since we can now fetch and clone shallow repositories in
interoperability mode, add support for pushing as well.  If we have
shallow data, send an object-format-map capability that indicates that
the client will send object mappings before the shallows, both for the
shallows themselves and their trees and parents, so that the index-pack
operation can properly map them.

We send the mappings in this order to ensure that we can process data in
one pass, which is more efficient than buffering and waiting to handle
the shallows until later.  It will also allow us to gracefully handle
submodules and partial clones in a single pass as well in the future.

Control the sending of mapping with a configuration option so that
people who do not want to expose the use of compatibility mode can
disable it.  However, default it to on since it provides important
functionality and our testsuite is much more functional in compatibility
mode with it on.

As with fetches, make sure to use only index-pack to handle data when
the receiving system has a compatibility mode as the unpack-objects code
path is not updated to handle mapping objects and it will be less
efficient to use the loose object format anyway.