t6020: match bundle correctly in compatibility mode
When we're running the testsuite in compatibility mode, the bundle will
have only a single hash algorithm and so the expected message will not
match GIT_DEFAULT_HASH, which will have two algorithms separated by a
colon.  Fix the message to look for only the main algorithm to fix
several assertions in this mode.