upload-pack: split packing arg computation from actual packing
Our code here to invoke pack-objects makes a lot of decisions about
which revisions to include and which to exclude.  In a future commit,
we'll need to perform similar functionality to iterate over trees to
find any submodules and we'll want to minimize duplication.

Split out the computation of the packing arguments into a function that
is separate from the actual invocation of pack-objects.  For now, there
is no functional change and we collect no additional data, but in the
future, we'll compute additional arguments for our internal rev-list and
add them to struct upload_pack_data for use elsewhere in upload-pack.

Signed-off-by: brian m. carlson <bk2204@github.com>