oidtree: allow querying whether tree is empty
When we are emitting object mappings as part of the protocol, we'd like
to omit the header and footer, as well as the expensive portions of
traversing data structures, if there are no mappings to emit.

In a future commit, we'll use an oidtree to store the object IDs of the
submodules we need to emit because this provides quick, easy
de-duplication as well as in-order traversal.  However, if there are no
entries, then we may end up skipping the mention of any object mappings.
To make this possible to determine, add a helper that queries whether a
tree is empty or not.

Signed-off-by: brian m. carlson <bk2204@github.com>